Add files via upload
This program describes about  Best time to buy and sell stock. You are given an array prices where prices[i] is the price of a given stock on the ith day. You want to maximize your profit by choosing a single day choosing a different day in the future to sell that stock. to buy one stock and choosing a different day in the future to sell that stock. Return the maximum profit you can achieve from this transaction. if you cannot achieve any profit, return 0.